+++ elasticsearch2/src/test/java/org/testifyproject/resource/elasticsearch/Elasticsearch2ResourceTest.java
@@ -26,7 +26,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
-import org.testifyproject.ResourceInstance;
+import org.testifyproject.LocalResourceInstance;
 import org.testifyproject.TestContext;
 import org.testifyproject.annotation.Cut;
 import org.testifyproject.annotation.Fixture;
@@ -51,7 +51,7 @@ public class Elasticsearch2ResourceTest {
         Settings.Builder config = cut.configure(testContext);
         assertThat(config).isNotNull();
 
-        ResourceInstance<Node, Client> result = cut.start(testContext, config);
+        LocalResourceInstance<Node, Client> result = cut.start(testContext, config);
 
         assertThat(result).isNotNull();
         assertThat(result.getClient()).isPresent();