+++ hdfs/src/test/java/org/testifyproject/resource/hdfs/MiniDFSResourceTest.java
@@ -28,7 +28,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
-import org.testifyproject.ResourceInstance;
+import org.testifyproject.LocalResourceInstance;
 import org.testifyproject.TestContext;
 import org.testifyproject.annotation.Cut;
 import org.testifyproject.annotation.Fixture;
@@ -53,7 +53,7 @@ public class MiniDFSResourceTest {
         HdfsConfiguration config = cut.configure(testContext);
         assertThat(config).isNotNull();
 
-        ResourceInstance<MiniDFSCluster, DistributedFileSystem> result = cut.start(testContext, config);
+        LocalResourceInstance<MiniDFSCluster, DistributedFileSystem> result = cut.start(testContext, config);
 
         assertThat(result).isNotNull();
         assertThat(result.getClient()).isPresent();