+++ storm/src/test/java/org/testifyproject/resource/storm/StormResourceTest.java
@@ -28,7 +28,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
-import org.testifyproject.ResourceInstance;
+import org.testifyproject.LocalResourceInstance;
 import org.testifyproject.TestContext;
 import org.testifyproject.annotation.Cut;
 import org.testifyproject.annotation.Fixture;
@@ -55,7 +55,7 @@ public class StormResourceTest {
         Void config = cut.configure(testContext);
         assertThat(config).isNull();
 
-        ResourceInstance<ILocalCluster, Void> result = cut.start(testContext, config);
+        LocalResourceInstance<ILocalCluster, Void> result = cut.start(testContext, config);
 
         assertThat(result).isNotNull();
         assertThat(result.getClient()).isEmpty();