+++ titan/src/test/java/org/testifyproject/resource/titan/TitanBerkeleyResourceTest.java
@@ -23,7 +23,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
-import org.testifyproject.ResourceInstance;
+import org.testifyproject.LocalResourceInstance;
 import org.testifyproject.TestContext;
 import org.testifyproject.annotation.Cut;
 import org.testifyproject.annotation.Fixture;
@@ -48,7 +48,7 @@ public class TitanBerkeleyResourceTest {
         CommonsConfiguration config = cut.configure(testContext);
         assertThat(config).isNotNull();
 
-        ResourceInstance<TitanGraph, GraphTraversalSource> result = cut.start(testContext, config);
+        LocalResourceInstance<TitanGraph, GraphTraversalSource> result = cut.start(testContext, config);
 
         assertThat(result).isNotNull();
         assertThat(result.getClient()).isNotEmpty();